Address @pshriwise comments on #3119
