+++ include/openmc/surface.h
@@ -9,6 +9,7 @@
 #include "pugixml.hpp"
 
 #include "openmc/boundary_condition.h"
+#include "openmc/bounding_box.h"
 #include "openmc/constants.h"
 #include "openmc/memory.h" // for unique_ptr
 #include "openmc/particle.h"
@@ -27,55 +28,6 @@ namespace model {
 extern std::unordered_map<int, int> surface_map;
 extern vector<unique_ptr<Surface>> surfaces;
 } // namespace model
-
-//==============================================================================
-//! Coordinates for an axis-aligned cuboid that bounds a geometric object.
-//==============================================================================
-
-struct BoundingBox {
-  double xmin = -INFTY;
-  double xmax = INFTY;
-  double ymin = -INFTY;
-  double ymax = INFTY;
-  double zmin = -INFTY;
-  double zmax = INFTY;
-
-  inline BoundingBox operator&(const BoundingBox& other)
-  {
-    BoundingBox result = *this;
-    return result &= other;
-  }
-
-  inline BoundingBox operator|(const BoundingBox& other)
-  {
-    BoundingBox result = *this;
-    return result |= other;
-  }
-
-  // intersect operator
-  inline BoundingBox& operator&=(const BoundingBox& other)
-  {
-    xmin = std::max(xmin, other.xmin);
-    xmax = std::min(xmax, other.xmax);
-    ymin = std::max(ymin, other.ymin);
-    ymax = std::min(ymax, other.ymax);
-    zmin = std::max(zmin, other.zmin);
-    zmax = std::min(zmax, other.zmax);
-    return *this;
-  }
-
-  // union operator
-  inline BoundingBox& operator|=(const BoundingBox& other)
-  {
-    xmin = std::min(xmin, other.xmin);
-    xmax = std::max(xmax, other.xmax);
-    ymin = std::min(ymin, other.ymin);
-    ymax = std::max(ymax, other.ymax);
-    zmin = std::min(zmin, other.zmin);
-    zmax = std::max(zmax, other.zmax);
-    return *this;
-  }
-};
 
 //==============================================================================
 //! A geometry primitive used to define regions of 3D space.
